requirements.txt
aiogram<3.0
environs~=8.0.0
tortoise-orm==0.19.0
asyncpg==0.25.0
aerich==0.6.2
black==22.10.0